AK: Move buffered streams from `LibCore`
